Ownership for update and delete

+++ Tham-Backend/Controllers/CommentLikesController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Tham_Backend.Models;
 using Tham_Backend.Repositories;
+using Tham_Backend.Services;
 
 namespace Tham_Backend.Controllers;
 
@@ -10,10 +11,12 @@ namespace Tham_Backend.Controllers;
 public class CommentLikesController : ControllerBase
 {
     private readonly ICommentLikeRepository _repository;
+    private readonly IUserService _userService;
 
-    public CommentLikesController(ICommentLikeRepository repository)
+    public CommentLikesController(ICommentLikeRepository repository, IUserService userService)
     {
         _repository = repository;
+        _userService = userService;
     }
 
     [HttpGet]
@@ -45,17 +48,44 @@ public class CommentLikesController : ControllerBase
 
     [HttpPut("{id:min(1)}")]
     [Authorize(Roles = "Admin,User")]
-    public async Task<IActionResult> UpdateCommentLike([FromRoute] int id, [FromBody] CommentLikeModel commentLikeModel)
+    public async Task<IActionResult> UpdateCommentLike([FromRoute] int id, [FromBody] CommentLikeModel commentLikeModel, [FromServices]IBloggerRepository bloggerRepository)
     {
-        await _repository.UpdateCommentLikeAsync(id, commentLikeModel);
-        return Ok();
+        var oldCommentLike = await _repository.GetCommentLikeByIdAsync(id);
+        if (oldCommentLike is not null)
+        { 
+            var user = await bloggerRepository.GetBloggerByIdAsync((int) oldCommentLike.BloggerId);
+            if (user is not null)
+            {
+                if (_userService.GetRole() == "User")
+                {
+                    if (_userService.GetEmail() != user.Email) return Unauthorized("You are not owner of this comment like!");
+                }
+                await _repository.UpdateCommentLikeAsync(id, commentLikeModel);
+                return Ok();
+            }
+        }
+
+        return NotFound();
     }
 
     [HttpDelete("{id:min(1)}")]
     [Authorize(Roles = "Admin,User")]
-    public async Task<IActionResult> DeleteCommentLike([FromRoute] int id)
-    {
-        await _repository.DeleteCommentLikeAsync(id);
-        return Ok();
+    public async Task<IActionResult> DeleteCommentLike([FromRoute] int id,[FromServices]IBloggerRepository bloggerRepository)
+    {var oldCommentLike = await _repository.GetCommentLikeByIdAsync(id);
+        if (oldCommentLike is not null)
+        { 
+            var user = await bloggerRepository.GetBloggerByIdAsync((int) oldCommentLike.BloggerId);
+            if (user is not null)
+            {
+                if (_userService.GetRole() == "User")
+                {
+                    if (_userService.GetEmail() != user.Email) return Unauthorized("You are not owner of this comment like!");
+                }
+                await _repository.DeleteCommentLikeAsync(id);
+                return Ok();
+            }
+        }
+        
+        return NotFound();
     }
 }
